refactor: Organize interfaces into separate files
- Add IFlowRepository with simple CRUD methods (saveFlow, getFlow, listFlows, deleteFlow)
- Add IDomainFlowRepository for future domain-driven design
- Split interfaces into separate files for better organization:
  - IFlowRepository.ts: Flow repository + domain entities
  - IUtils.ts: IFeatureManager, ILogger utility interfaces
  - ILegacy.ts: Deprecated IAuthProvider interface
- Keep index.ts as clean re-exports only
